accounting_assistant/requirements.txt
schedule==1.2.0
requests==2.31.0
openai==1.3.0
pandas==2.1.0
python-dateutil==2.8.2
smtplib-ssl==1.0.4
python-dotenv==1.0.0
sqlite3
logging
